+++ seq_to_first_iso/__init__.py
@@ -1,12 +1,13 @@
 
-"""Compute first two isotopologue intensities from sequences.
+"""Compute intensities of the first two isotopologue 
+from peptide sequences and charges.
 
 The program computes M0 and M1 and differentiate labelled
 (with a 99.99 % C[12] enrichment) and unlabelled amino acids.
 
-Read a file composed of amino acid sequences on each line and return :
+Read a .tsv file composed of amino acid sequences on each line and return:
 sequence, mass, formula, formula_X, M0_NC, M1_NC, M0_12C and M1_12C
-in a tsv file.
+in a .tsv file.
 
 Formula_X is the chemical formula with carbon of unlabelled
 amino acids marked as X.
@@ -19,7 +20,7 @@ Example
 
 Running the script after installation
 
-    $ seq-to-first-iso sequences.txt
+    $ seq-to-first-iso sequences.tsv sequence_column_name charge_column_name
 
 will provide file 'sequences_stfi.tsv'
 
@@ -34,33 +35,39 @@ Naming conventions for isotopes follow pyteomics's conventions.
 
 __authors__ = "Lilian Yang-crosson, Pierre Poulain"
 __license__ = "BSD 3-Clause License"
-__version__ = "0.5.1"
+__version__ = "1.0.0"
 __maintainer__ = "Pierre Poulain"
 __email__ = "pierre.poulain@cupnet.net"
 
 from .seq_to_first_iso import (AMINO_ACIDS,
-                               C12_abundance,
-                               isotopic_abundance,
+                               XTANDEM_MOD_PATTERN,
                                UNIMOD_MODS,
-                               sequence_parser,
+                               normal_abundance,
+                               C12_abundance,
+                               parse_input_file,
+                               check_amino_acids,
                                separate_labelled,
                                compute_M0_nl,
                                compute_M1_nl,
                                formula_to_str,
-                               seq_to_xcomp,
+                               convert_atom_C_to_X,
+                               get_charge_composition,
                                get_mods_composition,
-                               seq_to_df,
+                               compute_intensities,
                                )
 __all__ = ["AMINO_ACIDS",
-           "C12_abundance",
-           "isotopic_abundance",
+           "XTANDEM_MOD_PATTERN",
            "UNIMOD_MODS",
-           "sequence_parser",
+           "normal_abundance",
+           "C12_abundance",
+           "parse_input_file",
+           "check_amino_acids",
            "separate_labelled",
            "compute_M0_nl",
            "compute_M1_nl",
            "formula_to_str",
-           "seq_to_xcomp",
+           "convert_atom_C_to_X",
+           "get_charge_composition",
            "get_mods_composition",
-           "seq_to_df",
+           "compute_intensities",
            ]